feat!: Support fractional fonts and grid sizes (try 2) (#2500)
* Always use fractional font sizes

NOTE: The grid is still integer sized, so it's rendered wrong

* Remove fudge factor

* Properly set a fractional grid size

* Disable baseline snapping, and enable rounding to integers for scroll offsets again

The baseline snapping needs to be disabled so that the lines are
properly placed, and not snapped to pixels. While using integer based
scroll offsets greatly reduces the pressure on the Skia font cache.

* Plot the skia cache sizes

* Improve the Skia font cache usage

A bigger default size with cleanup when idling

* Fix fractional border positions

* Update the documentation

* Enable baseline snap

* Ceil the font height to the nearest pixel

This improves the quality and the performance of the font rendering.

+++ src/editor/mod.rs
@@ -554,7 +554,7 @@ impl Editor {
             }
             GuiOption::LineSpace(linespace) => {
                 self.draw_command_batcher
-                    .queue(DrawCommand::LineSpaceChanged(linespace));
+                    .queue(DrawCommand::LineSpaceChanged(linespace as f32));
 
                 self.redraw_screen();
             }
